feat: Add solution for LeetCode problem 19 - Remove Nth Node From End of List
This commit adds a C++ solution for LeetCode problem 19, \"Remove Nth Node From End of List\". The solution uses the two-pointer approach (fast and slow pointers) to achieve a single-pass removal of the nth node from the end of the linked list. A dummy node is used to handle edge cases like removing the head node.